Rename rock-paper-scissors-determine-winner exercise to rock-paper-scissors
Shorten the slug and title to rock-paper-scissors / "Rock, Paper,
Scissors". Renames the exercise directory and lesson icon, and updates
the registry, llm-metadata registry, instructions title, metadata slug,
and the README/PLAN references.
