+++ curriculum/src/exercises/index.ts
@@ -65,7 +65,7 @@ export const exercises = {
   "cityscape-skyline": () => import("./cityscape-skyline"),
   "rainbow-splodges": () => import("./rainbow-splodges"),
   "digital-clock": () => import("./digital-clock"),
-  "rock-paper-scissors-determine-winner": () => import("./rock-paper-scissors-determine-winner"),
+  "rock-paper-scissors": () => import("./rock-paper-scissors"),
   "maze-automated-solve": () => import("./maze-automated-solve"),
   "maze-turn-around": () => import("./maze-turn-around"),
   "rainbow-ball": () => import("./rainbow-ball"),
