+++ curriculum/src/exercises/rock-paper-scissors/Exercise.ts
