+++ curriculum/src/exercises/rock-paper-scissors/index.ts
