+++ curriculum/src/exercises/rock-paper-scissors/llm-metadata.ts
