+++ curriculum/src/exercises/rock-paper-scissors/metadata.json
@@ -1,5 +1,5 @@
 {
-  "slug": "rock-paper-scissors-determine-winner",
+  "slug": "rock-paper-scissors",
   "levelId": "complex-conditionals",
   "hints": [
     {
