+++ curriculum/src/exercises/rock-paper-scissors/scenarios.ts
