+++ curriculum/src/exercises/rock-paper-scissors/solution.javascript
