+++ curriculum/src/exercises/rock-paper-scissors/solution.jiki
