+++ curriculum/src/exercises/rock-paper-scissors/stub.javascript
