+++ curriculum/src/exercises/rock-paper-scissors/stub.jiki
