+++ curriculum/src/llm-metadata.ts
@@ -71,7 +71,7 @@ import { llmMetadata as cityscapeSkyscraperLLM } from "./exercises/cityscape-sky
 import { llmMetadata as cityscapeSkylineLLM } from "./exercises/cityscape-skyline/llm-metadata";
 import { llmMetadata as rainbowSplodgesLLM } from "./exercises/rainbow-splodges/llm-metadata";
 import { llmMetadata as digitalClockLLM } from "./exercises/digital-clock/llm-metadata";
-import { llmMetadata as rockPaperScissorsDetermineWinnerLLM } from "./exercises/rock-paper-scissors-determine-winner/llm-metadata";
+import { llmMetadata as rockPaperScissorsLLM } from "./exercises/rock-paper-scissors/llm-metadata";
 import { llmMetadata as mazeAutomatedSolveLLM } from "./exercises/maze-automated-solve/llm-metadata";
 import { llmMetadata as mazeTurnAroundLLM } from "./exercises/maze-turn-around/llm-metadata";
 import { llmMetadata as rainbowBallLLM } from "./exercises/rainbow-ball/llm-metadata";
@@ -183,7 +183,7 @@ const llmMetadataRegistry = {
   "cityscape-skyline": cityscapeSkylineLLM,
   "rainbow-splodges": rainbowSplodgesLLM,
   "digital-clock": digitalClockLLM,
-  "rock-paper-scissors-determine-winner": rockPaperScissorsDetermineWinnerLLM,
+  "rock-paper-scissors": rockPaperScissorsLLM,
   "maze-automated-solve": mazeAutomatedSolveLLM,
   "maze-turn-around": mazeTurnAroundLLM,
   "rainbow-ball": rainbowBallLLM,
